+++ CHANGELOG.md
@@ -25,6 +25,7 @@ That's all, matcher will stop when detects next line started with `###` match
 * Fixed restoring door blocks on legacy versions
 * Changed empty message is no longer sent
 * Added Thai and Turkish locale (Thanks to PoEditor Contributors)
+* Added configurable name visibility for Golem, Wolf and Villager
 
 ### 4.5.9 (28.03.2021 - 13.04.2021)
 * Fix when Splash_potion material not existed on 1.8 servers